refactor: Move loading RPC URLs into a reusable hook.
